qemuloongarch64: Initial support for QEMU LoongArch64
LoongArch is a new ISA introduced by Loongson[1], supported since linux-5.19
and continuously maintained.

Starting in 2023, oe-core adds support for LoongArch. In order to generate
.config files using .scc and .cfg when building the kernel, yocto-kernel-cache
also needs to add support for the LoongArch architecture.

The patch provides a basic configuration and passes the boot test using
qemu-system-loongarch64.

The meta-loongarch[2] maintenance team is responsible for this repository's
LoongArch architecture.

[1]: https://loongson.cn/EN
[2]: https://layers.openembedded.org/layerindex/branch/nanbield/layer/meta-loongarch/

Signed-off-by: Xiaotian Wu <wuxiaotian@loongson.cn>